Separating logic for readability
